+++ yash-env/src/trap.rs
@@ -945,10 +945,7 @@ mod tests {
                     false,
                 )
                 .unwrap();
-            env.system
-                .kill(env.main_pid, Some(Signal::SIGINT))
-                .await
-                .unwrap();
+            env.system.kill(env.main_pid, Some(SIGINT)).await.unwrap();
             env.traps.enter_subshell(&mut env.system, true, false);
 
             let state = state.borrow();
@@ -970,10 +967,7 @@ mod tests {
                     false,
                 )
                 .unwrap();
-            env.system
-                .kill(env.main_pid, Some(Signal::SIGQUIT))
-                .await
-                .unwrap();
+            env.system.kill(env.main_pid, Some(SIGQUIT)).await.unwrap();
             env.traps.enter_subshell(&mut env.system, true, false);
 
             let state = state.borrow();
@@ -1007,7 +1001,7 @@ mod tests {
                     .unwrap();
             }
             env.traps.enable_stopper_handlers(&mut env.system).unwrap();
-            for signal in [Signal::SIGTSTP, Signal::SIGTTIN, Signal::SIGTTOU] {
+            for signal in [SIGTSTP, SIGTTIN, SIGTTOU] {
                 env.system.kill(env.main_pid, Some(signal)).await.unwrap();
             }
             env.traps.enter_subshell(&mut env.system, false, true);